files paths refactored good-form component added
