+++ ShopForWood/app/components/goods/components/edit-good/edit-good.component.ts
@@ -1,19 +1,19 @@
 import { Component, OnInit, Input, OnDestroy, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
 import { Params, ActivatedRoute, Router } from '@angular/router';
-import { Subscription } from 'rxjs/Subscription';
+import { Good } from '../../models/good.model';
 import { GoodService } from '../../services/good.service';
-import { Good } from '../good/models/good.model';
+import { Subscription } from 'rxjs/Subscription';
 import 'rxjs/add/operator/mergeMap';
 
 @Component({
     moduleId: module.id,
     selector: 'edit-good',
     template: `
-        <div *ngIf="good !== undefined">
-            <input type="text" [(ngModel)]="good.name"/>
-            <input type="text" [(ngModel)]="good.description"/>
-        </div>
-        <button (click)="Save()">Save</button>
+        <good-form 
+            [good]="good"
+            [submitButtonName]="submitButtonName"
+            (submit)="Save()">
+        </good-form>
     `,
     styleUrls: ['./edit-good.component.css'],
     changeDetection: ChangeDetectionStrategy.OnPush
@@ -22,7 +22,8 @@ import 'rxjs/add/operator/mergeMap';
 export class EditGoodComponent implements OnInit, OnDestroy {
     private goodId: number;
     private subscription: Subscription;
-    public good: Good;
+    private good: Good;
+    private submitButtonName: string = "Edit";
 
     constructor(
         private _goodService: GoodService,
