+++ ShopForWood/app/components/goods/components/edit-good/edit-good.module.ts
@@ -3,11 +3,12 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { EditGoodComponent } from './edit-good.component';
 import { GoodService } from '../../services/good.service';
+import { GoodFormComponent } from '../good-form/good-form.component';
 
 @NgModule({
     imports: [CommonModule, FormsModule],
     exports: [EditGoodComponent],
-    declarations: [EditGoodComponent],
+    declarations: [EditGoodComponent, GoodFormComponent],
     providers: [GoodService],
 })
 export class EditGoodModule { }
