+++ ShopForWood/app/components/goods/components/good-form/good-form.component.ts
@@ -1,0 +1,34 @@
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Response } from '@angular/http';
+import { Router } from '@angular/router';
+import { Good } from '../../models/good.model';
+
+@Component({
+    moduleId: module.id,
+    selector: 'good-form',
+    template: `
+        <div *ngIf="good !== undefined">
+            <div>
+                <input type="text" [(ngModel)]="good.name"/>
+            </div>
+            <div>
+                <input type="text" [(ngModel)]="good.description"/>
+            </div>
+            <button (click)="SubmitFrom()">{{submitButtonName}}</button>
+        </div>
+    `,
+    styleUrls: ['./good-form.component.css']
+})
+export class GoodFormComponent implements OnInit {
+    @Input() good: Good = new Good();
+    @Input() submitButtonName: string;
+    @Output() submit: EventEmitter<Good> = new EventEmitter<Good>();
+
+    constructor(private _router: Router) { }
+
+    ngOnInit() { }
+
+    private SubmitFrom() {
+        this.submit.emit(this.good);
+    }
+}
