+++ ShopForWood/app/components/goods/models/good.model.ts
