+++ ShopForWood/app/components/goods/services/good.service.ts
@@ -3,7 +3,7 @@ import { Http, Response } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
-import { Good } from '../components/good/models/good.model';
+import { Good } from '../models/good.model';
 
 @Injectable()
 export class GoodService {
